added profile picture support
