+++ app/src/main/java/com/example/nutritionapp/Login.java
@@ -8,7 +8,6 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
-import com.example.nutritionapp.ui.profile.ProfileFragment;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
@@ -75,7 +74,7 @@ public class Login extends AppCompatActivity {
             dm.setEmail(googleSignInAccount1.getEmail());
             dm.setImageUrl(googleSignInAccount1.getPhotoUrl());
             dm.setName2(googleSignInAccount1.getGivenName());
-            Intent intent=new Intent(this, Profile.class);
+            Intent intent=new Intent(this, MyProfile.class);
             intent.putExtra("obj",dm);
             startActivity(intent);
         }
