+++ app/src/main/res/layout/activity_my_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Profile">
+    tools:context=".MyProfile">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -13,8 +13,10 @@
         android:padding="30dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="1.0">
 
         <ImageView
             android:id="@+id/google_imgw"
@@ -52,6 +54,27 @@
             android:textSize="18sp" />
 
     </LinearLayout>
+    <androidx.cardview.widget.CardView
+    android:id="@+id/cardView"
+    android:layout_width="180dp"
+    android:layout_height="180dp"
+    app:cardCornerRadius="90dp"
+    app:layout_constraintVertical_bias="0.4"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent">
+
+
+        <ImageView
+            android:id="@+id/profile_picture"
+            android:layout_width="match_parent"
+            android:layout_height="184dp"
+            android:scaleType="centerCrop"
+            android:src="@drawable/ic_profile_photo" />
+
+
+</androidx.cardview.widget.CardView>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
